chatbot/pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Manuele Lamenha <manuele.lamenha.mauricio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
black = "^24.10.0"
pytest = "^8.3.3"
mkdocs = "^1.6.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
markdown = "^3.7"
django-rest-framework = "^0.1.0"
pytelegrambotapi = "^4.24.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
autoflake = "^2.3.1"
discord-py = "^2.4.0"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.8"
pytest-django = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
include = '\.piy?$'
exclude = '''
/(
  \.git
 | \.hg
 | \.mypy_cache
 | \.tox
 | \.venv
 | _build
 | buck-out
 | build
 )/
'''